use prompter and remove completions
